Move AccessClass and arch normalization into rise-backend-core
Both are consumed by code that is about to leave the root crate, and both
are genuinely shared rather than Kubernetes-specific, so core is where they
belong — putting them anywhere else would make the extracted backend depend
back on the root crate.

AccessClass is referenced by all three backends through their access_classes
maps. Its `custom_annotations` doc no longer says "nginx": Kubernetes stamps
nginx ingress annotations, but Docker and ECS consume the same type through
Traefik forwardAuth, and the schema description now says so. That is the only
change in the regenerated backend-settings schema.

normalize_runtime_arch sits next to BackendCapabilities, which is what
surfaces the value it produces.
